Time Range Data widget
Adds the Time Range Data Widget to the sidebar as specified in
ADR 0008: Time Range Data Widget

The unit controller is linked to the widget by signals. When a new
active tab is loaded or changed, the trace-viewer Theia component
dispatches the new active unit controller via signal-manager. This
signal is picked up by the time-range-data-widget react component.
Because of this, the unit controller is now a public value in the
trace-context-component.

Signed-off-by: William Yang <william.yang@ericsson.com>